Move `valkey_asprintf` to util.c
